+++ src/energyMonitoringInterface/energyMonitoringInterface.c
@@ -1,4 +1,4 @@
-#include "sub-system.h"
+#include "energyMonitoringInterface.h"
 #include <stdio.h>
 #include <stdint.h>
 #include <compat/twi.h>
